+++ services/notification/internal/core/hub.go
@@ -1,4 +1,4 @@
-package api
+package notification
 
 import (
 	"log"
